Split orchestration property tests (#72)
Guard finish callbacks so callback failures are logged without replacing
the already computed graph result.

Split the orchestration property checks into focused modules with shared
property helpers. Restore the model-tier boundary check to deterministic
parametrisation.
